Swapon with maximum priority before hibernation
With recent versions of systemd, on systems with additional or
multiple swapfiles configured, swapon on /swap will be added with the
next lowest priority for the system.

For example, if we have /swapfile configured with a default priority
of -2, swapon /swap will be added with priority -3.

$ sudo swapon /swapfile
$ swapon --show
NAME      TYPE SIZE USED PRIO
/swapfile file   4G   0B   -2
$ sudo swapon /swap
$ swapon --show
NAME          TYPE SIZE USED PRIO
/swapfile     file   4G   0B   -2
/swap         file 3.9G   0B   -3

This causes hibernation to fail, since hibernation targets the
highest priority swapfile, which is no longer /swap, and the
kernel command line is specifically configured to resume from the
UUID and offset for /swap.

The result is that EC2 console says "Stopping" for 20 minutes before
the instance is force stopped.

Fix this by setting /swap to maximum possible priority,
32767, as we swapon right before hibernation takes place.